Updated home component
