+++ src/components/content/userinfo/personalinfo/EditUserData.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Container, Form } from "react-bootstrap";
+import { Container } from "react-bootstrap";
 import UserDataForm from "./UserDataForm";
 
 const EditUserData = () => {
