+++ src/components/content/userinfo/trustedperson/EditTrustedPersonData.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Container, Form } from "react-bootstrap";
+import { Container } from "react-bootstrap";
 import TrustedPersonForm from "./TrustedPersonForm";
 
 const EditTrustedPersonData = () => {
